all correct

+++ Libft/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: dasantos <dasantos@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 12:04:32 by dasantos          #+#    #+#             */
-/*   Updated: 2025/10/14 13:30:20 by dasantos         ###   ########.fr       */
+/*   Updated: 2025/10/16 12:34:24 by dasantos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-void ft_bzero(void *s, size_t n)
+void ft_bzero(void *s, unsigned int n)
 {
     ft_memset(s, 0, n);
 }
