+++ Libft/ft_strlcpy.c
@@ -6,18 +6,18 @@
 /*   By: dasantos <dasantos@student.42porto.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/14 14:30:26 by dasantos          #+#    #+#             */
-/*   Updated: 2025/10/14 14:37:21 by dasantos         ###   ########.fr       */
+/*   Updated: 2025/10/16 12:29:10 by dasantos         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-size_t ft_strlcpy(char *dst, const char *src, size_t dstsize)
+unsigned int ft_strlcpy(char *dst, const char *src, unsigned int dstsize)
 {
-    size_t srclen = ft_strlen(src);
+    int srclen = ft_strlen(src);
     if (dstsize == 0)
         return (srclen);
-    size_t i = 0;
+    int i = 0;
     while (i + 1 < dstsize && src[i])
     {
         dst[i] = src[i];
